even more tests, ready to merge
