+++ memory_addr/src/lib.rs
@@ -77,72 +77,20 @@ pub const fn is_aligned_4k(addr: usize) -> bool {
 
 #[cfg(test)]
 mod tests {
-    use crate::{va, va_range, MemoryAddr, VirtAddrRange};
+    use super::*;
 
     #[test]
-    fn test_addr() {
-        let addr = va!(0x2000);
-        assert!(addr.is_aligned_4k());
-        assert!(!addr.is_aligned(0x10000usize));
-        assert_eq!(addr.align_offset_4k(), 0);
-        assert_eq!(addr.align_down_4k(), va!(0x2000));
-        assert_eq!(addr.align_up_4k(), va!(0x2000));
+    fn test_align() {
+        assert_eq!(align_down(0x12345678, 0x1000), 0x12345000);
+        assert_eq!(align_up(0x12345678, 0x1000), 0x12346000);
+        assert_eq!(align_offset(0x12345678, 0x1000), 0x678);
+        assert!(is_aligned(0x12345000, 0x1000));
+        assert!(!is_aligned(0x12345678, 0x1000));
 
-        let addr = va!(0x2fff);
-        assert!(!addr.is_aligned_4k());
-        assert_eq!(addr.align_offset_4k(), 0xfff);
-        assert_eq!(addr.align_down_4k(), va!(0x2000));
-        assert_eq!(addr.align_up_4k(), va!(0x3000));
-
-        let align = 0x100000;
-        let addr = va!(align * 5) + 0x2000;
-        assert!(addr.is_aligned_4k());
-        assert!(!addr.is_aligned(align));
-        assert_eq!(addr.align_offset(align), 0x2000);
-        assert_eq!(addr.align_down(align), va!(align * 5));
-        assert_eq!(addr.align_up(align), va!(align * 6));
-    }
-
-    #[test]
-    fn test_range() {
-        let start = va!(0x1000);
-        let end = va!(0x2000);
-        let range = va_range!(start..end);
-        println!("range: {:?}", range);
-
-        assert!((0x1000..0x1000).is_empty());
-        assert!((0x1000..0xfff).is_empty());
-        assert!(!range.is_empty());
-
-        assert_eq!(range.start, start);
-        assert_eq!(range.end, end);
-        assert_eq!(range.size(), 0x1000);
-
-        assert!(range.contains(va!(0x1000)));
-        assert!(range.contains(va!(0x1080)));
-        assert!(!range.contains(va!(0x2000)));
-
-        assert!(!range.contains_range((0xfff..0x1fff).into()));
-        assert!(!range.contains_range((0xfff..0x2000).into()));
-        assert!(!range.contains_range((0xfff..0x2001).into()));
-        assert!(range.contains_range((0x1000..0x1fff).into()));
-        assert!(range.contains_range((0x1000..0x2000).into()));
-        assert!(!range.contains_range((0x1000..0x2001).into()));
-        assert!(range.contains_range((0x1001..0x1fff).into()));
-        assert!(range.contains_range((0x1001..0x2000).into()));
-        assert!(!range.contains_range((0x1001..0x2001).into()));
-        assert!(!range.contains_range(VirtAddrRange::from_start_size(0xfff.into(), 0x1)));
-        assert!(!range.contains_range(VirtAddrRange::from_start_size(0x2000.into(), 0x1)));
-
-        assert!(range.contained_in((0xfff..0x2000).into()));
-        assert!(range.contained_in((0x1000..0x2000).into()));
-        assert!(range.contained_in((0x1000..0x2001).into()));
-
-        assert!(!range.overlaps((0x800..0x1000).into()));
-        assert!(range.overlaps((0x800..0x1001).into()));
-        assert!(range.overlaps((0x1800..0x2000).into()));
-        assert!(range.overlaps((0x1800..0x2001).into()));
-        assert!(!range.overlaps((0x2000..0x2800).into()));
-        assert!(range.overlaps((0xfff..0x2001).into()));
+        assert_eq!(align_down_4k(0x12345678), 0x12345000);
+        assert_eq!(align_up_4k(0x12345678), 0x12346000);
+        assert_eq!(align_offset_4k(0x12345678), 0x678);
+        assert!(is_aligned_4k(0x12345000));
+        assert!(!is_aligned_4k(0x12345678));
     }
 }
