+++ memory_addr/src/range.rs
@@ -314,3 +314,67 @@ macro_rules! pa_range {
         $crate::PhysAddrRange::from($range)
     };
 }
+
+#[cfg(test)]
+mod test {
+    use crate::{va, va_range, VirtAddrRange};
+
+    #[test]
+    fn test_range_format() {
+        let range = va_range!(0xfec000usize..0xfff000usize);
+
+        assert_eq!(format!("{:?}", range), "VA:0xfec000..VA:0xfff000");
+        assert_eq!(format!("{:x}", range), "VA:0xfec000..VA:0xfff000");
+        assert_eq!(format!("{:X}", range), "VA:0xFEC000..VA:0xFFF000");
+    }
+
+    #[test]
+    fn test_range() {
+        let start = va!(0x1000);
+        let end = va!(0x2000);
+        let range = va_range!(start..end);
+
+        println!("range: {:?}", range);
+
+        assert!((0x1000..0x1000).is_empty());
+        assert!((0x1000..0xfff).is_empty());
+        assert!(!range.is_empty());
+
+        assert_eq!(range.start, start);
+        assert_eq!(range.end, end);
+        assert_eq!(range.size(), 0x1000);
+
+        assert!(range.contains(va!(0x1000)));
+        assert!(range.contains(va!(0x1080)));
+        assert!(!range.contains(va!(0x2000)));
+
+        assert!(!range.contains_range((0xfff..0x1fff).into()));
+        assert!(!range.contains_range((0xfff..0x2000).into()));
+        assert!(!range.contains_range((0xfff..0x2001).into()));
+        assert!(range.contains_range((0x1000..0x1fff).into()));
+        assert!(range.contains_range((0x1000..0x2000).into()));
+        assert!(!range.contains_range((0x1000..0x2001).into()));
+        assert!(range.contains_range((0x1001..0x1fff).into()));
+        assert!(range.contains_range((0x1001..0x2000).into()));
+        assert!(!range.contains_range((0x1001..0x2001).into()));
+        assert!(!range.contains_range(VirtAddrRange::from_start_size(0xfff.into(), 0x1)));
+        assert!(!range.contains_range(VirtAddrRange::from_start_size(0x2000.into(), 0x1)));
+
+        assert!(range.contained_in((0xfff..0x2000).into()));
+        assert!(range.contained_in((0x1000..0x2000).into()));
+        assert!(range.contained_in((0x1000..0x2001).into()));
+
+        assert!(!range.overlaps((0x800..0x1000).into()));
+        assert!(range.overlaps((0x800..0x1001).into()));
+        assert!(range.overlaps((0x1800..0x2000).into()));
+        assert!(range.overlaps((0x1800..0x2001).into()));
+        assert!(!range.overlaps((0x2000..0x2800).into()));
+        assert!(range.overlaps((0xfff..0x2001).into()));
+
+        let default_range: VirtAddrRange = Default::default();
+        assert!(default_range.is_empty());
+        assert_eq!(default_range.size(), 0);
+        assert_eq!(default_range.start, va!(0));
+        assert_eq!(default_range.end, va!(0));
+    }
+}
